requirements.txt
openai-whisper
torch
sounddevice
scipy
numpy
